+++ src/elements/Asset.php
@@ -990,7 +990,7 @@ class Asset extends Element
 
             list($value, $unit) = Assets::parseSrcsetSize($size);
 
-            $sizeTransform = [];
+            $sizeTransform = $transform->toArray();
             if ($unit === 'w') {
                 $sizeTransform['width'] = (int)$value;
             } else {
@@ -1004,10 +1004,6 @@ class Asset extends Element
                 } else {
                     $sizeTransform['height'] = (int)ceil($currentHeight * $value);
                 }
-            }
-
-            if (!empty($transform->format)) {
-                $sizeTransform['format'] = $transform->format;
             }
 
             $srcset[] = $this->getUrl($sizeTransform) . " $value$unit";